requirements.txt
fastapi==0.135.1
uvicorn==0.41.0
sqlalchemy==2.0.48
cryptography==46.0.5
pydantic== 2.12.5
httpx==0.28.1
asyncpg==0.31.0
python-dotenv==1.2.2
